Moved tq package to the root directory of tq
